# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Task Management application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Event": {
        "type": "object",
        "required": [
          "name",
          "date",
          "capacity"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the event",
            "example": "event_1"
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "description": "Name of the event",
            "example": "Graduation"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Must be a future date (greater than the current date and time)",
            "example": "2026-05-01T10:00:00Z"
          },
          "capacity": {
            "type": "integer",
            "minimum": 5,
            "description": "Must be a whole number greater than or equal to 5"
          },
          "registrationCount": {
            "type": "integer",
            "default": 0,
            "description": "Must be a whole number that is less than or equal to the capacity"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "completed"
            ],
            "default": "active"
          },
          "category": {
            "type": "string",
            "enum": [
              "conference",
              "workshop",
              "meetup",
              "seminar",
              "general"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-01T10:00:00Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type or code, followed by a readable error message"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/events": {
      "post": {
        "summary": "Create a new event",
        "tags": [
          "Events"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Event"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve a list of events",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "count": {
                      "type": "integer",
                      "description": "Number of events returned in the response"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Events retrieved successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "Retrieve an event by ID",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved event",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Event"
                    },
                    "message": {
                      "type": "string",
                      "example": "Event retrieved successfully"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing event",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "registrationCount": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "completed"
                    ]
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-05-01T10:00:00Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Event"
                    },
                    "message": {
                      "type": "string",
                      "example": "Event updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Registration count exceeds capacity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an existing event",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Event deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to delete the event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}